backend/requirements.txt
pydantic>=2.0.0
numpy>=1.24.0
networkx>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
jinja2>=3.0.0
tqdm>=4.65.0 
backoff>=2.2.0 